Add files via upload
Regarding the project that Farid and I presented last week. As per your request, we implemented the system using a vanilla CNN after initially building it with pretrained models like RexNet and MobileNet. We have completed the task and i send you now the code along with a screenshot showing the layers used in this implementation and the results ploted.

The vanilla CNN architecture we used is as follows:

Convolution Layers: 2 convolutional layers, the first with 16 filters and the second with 32 filters, both using a 3x3 kernel and ReLU activation.
Pooling: Max pooling layers (2x2) after each convolution to reduce spatial dimensions.
Fully Connected Layers: One fully connected layer with 64 neurons, followed by the output layer corresponding to the number of classes.

This simplified architecture was designed to intentionally reduce complexity to observe the contrast with pretrained models.
However, our results indicate that the pretrained models performed significantly better. This is likely due to their prior training on ImageNet, which enables them to handle complex tasks with greater efficiency and accuracy.

Zoalghinah Mohammad Fadi & HosseiniEbrahimAbad, SeyedFarid